show vessel historical positions
