+++ src/App.tsx
@@ -1,19 +1,22 @@
-import React from 'react';
+import React, {lazy, Suspense} from 'react';
 import './App.css';
 import 'antd/dist/antd.css';
-import List from "./features/vessel-track/list";
-import View from "./features/vessel-track/view";
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
+
+const List = lazy(() => import('./modules/vessel-track/list'));
+const View = lazy(() => import('./modules/vessel-track/view'));
 
 function App() {
     return (
         <div className="App">
-            <BrowserRouter>
-                <Routes>
-                    <Route path="/" element={<List/>}/>
-                    <Route path="/:id" element={<View/>}/>
-                </Routes>
-            </BrowserRouter>
+            <Suspense fallback="loading...">
+                <BrowserRouter>
+                    <Routes>
+                        <Route path="/" element={<List/>}/>
+                        <Route path="/:id" element={<View/>}/>
+                    </Routes>
+                </BrowserRouter>
+            </Suspense>
         </div>
     );
 }
